+++ CHANGELOG.md
@@ -13,6 +13,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0rc0] - 2023-06
+### Changed
+- Docstring inheritance for methods only inherit from the first found parent.
+### Fixed
+- Docstring inheritance for methods with no argument descriptions.
+- Format of the arguments provided with the default description for the Numpy style.
+
 ## [1.1.1] - 2023-01
 ### Fixed
 - Docstring inheritance for methods with multiple parents.